+++ dev/todo.md
@@ -1,0 +1,13 @@
+- action_set should take in linear classifier from scikit learn
+
+- add test script for cplex v pyomo
+    - should be easy to run  
+    - max cost is close
+    - no recourse cplex = no recourse pyomo;
+    - recourse cplex = recourse pyomo
+    - track cases where actions are not close and issue warning
+
+- adjusting the decision threshold 
+- return types for audit
+- recourse / cost 
+
